Update TaCQ for multi-gpu use-case

+++ SINQ/setting_perplexity_run.py
@@ -10,9 +10,11 @@ util_update_path = f"{os.path.join(os.path.dirname(lm_eval.__file__), f'models/v
 shutil.copyfile(util_source_path, util_update_path)
 
 # Multiblimp datasets
-shutil.copytree("../BHASA", "./BHASA")
-shutil.copytree("../syntactic_generalization_multilingual", "./syntactic_generalization_multilingual")
-shutil.copytree("../zhoblimp", "./zhoblimp")
+import os
+if not os.path.exists("./BHASA"):
+    shutil.copytree("../BHASA", "./BHASA")
+    shutil.copytree("../syntactic_generalization_multilingual", "./syntactic_generalization_multilingual")
+    shutil.copytree("../zhoblimp", "./zhoblimp")
 
 # True False answers for flips metric
 util_source_path = f"../edited_lm_eval/metrics.py"
